put collection into seperate users

+++ lib/app/features/home/pages/tasks/pages/category_pages/family_page/cubit/family_cubit.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 part 'family_state.dart';
 
@@ -9,7 +10,13 @@ class FamilyCubit extends Cubit<FamilyState> {
   FamilyCubit() : super(const FamilyState());
   StreamSubscription? _streamSubscription;
   Future<void> start() async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     _streamSubscription = FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('tasks')
         .where("category_id", isEqualTo: "eZhVFQmLWirJXdySFljK")
         .snapshots()
@@ -32,7 +39,13 @@ class FamilyCubit extends Cubit<FamilyState> {
 
   Future<void> remove({required String documentID}) async {
     try {
+      final userID = FirebaseAuth.instance.currentUser?.uid;
+      if (userID == null) {
+        throw Exception('User is not logged in');
+      }
       await FirebaseFirestore.instance
+          .collection('users')
+          .doc(userID)
           .collection('tasks')
           .doc(documentID)
           .delete();
