+++ lib/app/features/home/pages/tasks/pages/category_pages/house_page/cubit/house_cubit.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 part 'house_state.dart';
 
@@ -10,7 +11,13 @@ class HouseCubit extends Cubit<HouseState> {
 
   StreamSubscription? _streamSubscription;
   Future<void> start() async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     _streamSubscription = FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('tasks')
         .where("category_id", isEqualTo: "wRhqK5cwqBaQdrkPk6nT")
         .snapshots()
@@ -33,7 +40,13 @@ class HouseCubit extends Cubit<HouseState> {
 
   Future<void> remove({required String documentID}) async {
     try {
+      final userID = FirebaseAuth.instance.currentUser?.uid;
+      if (userID == null) {
+        throw Exception('User is not logged in');
+      }
       await FirebaseFirestore.instance
+          .collection('users')
+          .doc(userID)
           .collection('tasks')
           .doc(documentID)
           .delete();
